Move test fixtures to `src/fixtures`
