+++ crates/amalthea/src/fixtures/dummy_frontend.rs
