+++ crates/amalthea/src/lib.rs
@@ -8,6 +8,7 @@
 pub mod comm;
 pub mod connection_file;
 pub mod error;
+pub mod fixtures;
 pub mod kernel;
 pub mod kernel_dirs;
 pub mod kernel_spec;
@@ -16,7 +17,6 @@ pub mod session;
 pub mod socket;
 pub mod stream_capture;
 pub mod sys;
-pub mod test;
 pub mod wire;
 
 pub use error::Error;
