+++ crates/ark/src/lsp/completions/sources/composite.rs
@@ -153,7 +153,7 @@ mod tests {
     use crate::lsp::completions::sources::composite::is_identifier_like;
     use crate::lsp::document_context::DocumentContext;
     use crate::lsp::documents::Document;
-    use crate::test::r_test;
+    use crate::fixtures::r_test;
     use crate::treesitter::NodeType;
     use crate::treesitter::NodeTypeExt;
