+++ crates/ark/src/lsp/signature_help.rs
@@ -514,7 +514,7 @@ mod tests {
     use crate::lsp::documents::Document;
     use crate::lsp::signature_help::argument_label;
     use crate::lsp::signature_help::r_signature_help;
-    use crate::test::point_from_cursor;
+    use crate::fixtures::point_from_cursor;
 
     #[test]
     fn test_basic_signature_help() {
