+++ crates/ark/src/lsp/traits/node.rs
@@ -272,7 +272,7 @@ mod tests {
     use tree_sitter::Point;
 
     use crate::lsp::traits::node::NodeExt;
-    use crate::test::point_from_cursor;
+    use crate::fixtures::point_from_cursor;
 
     #[test]
     #[rustfmt::skip]
