+++ crates/ark/tests/connections.rs
@@ -14,8 +14,8 @@ use ark::connections::r_connection::Metadata;
 use ark::connections::r_connection::RConnection;
 use ark::modules::ARK_ENVS;
 use ark::r_task::r_task;
-use ark::test::r_test;
-use ark::test::socket_rpc_request;
+use ark::fixtures::r_test;
+use ark::fixtures::socket_rpc_request;
 use crossbeam::channel::bounded;
 use harp::exec::RFunction;
 use harp::object::RObject;
