+++ crates/ark/tests/data_explorer.rs
@@ -57,8 +57,8 @@ use ark::data_explorer::r_data_explorer::DataObjectEnvInfo;
 use ark::data_explorer::r_data_explorer::RDataExplorer;
 use ark::lsp::events::EVENTS;
 use ark::r_task::r_task;
-use ark::test::r_test;
-use ark::test::socket_rpc_request;
+use ark::fixtures::r_test;
+use ark::fixtures::socket_rpc_request;
 use ark::thread::RThreadSafe;
 use crossbeam::channel::bounded;
 use harp::environment::R_ENVS;
